# Parser optimizations + "Method code too large!" fix