--- a/server/grammar/score.bnf
+++ b/server/grammar/score.bnf
@@ -1,4 +1,4 @@
-score             = <ows> header? <ows> part* <ows>
+score             = <ows> header? part*
 header            = !calls <ows> music-data+
 part              = calls <ows> music-data*
 calls             = name