--- a/server/src/alda/lisp/events/barline.clj
+++ b/server/src/alda/lisp/events/barline.clj
@@ -1,6 +1,4 @@
 (ns alda.lisp.events.barline)
 
-(defn barline
-  "Barlines, at least currently, do nothing when evaluated in alda.lisp."
-  []
-  nil)
+(comment "Barlines, at least currently, do nothing when evaluated in alda.lisp.")
+