Closes #235, #238. Also helps with #208.

1. Parser optimizations by myself and @aengelberg -- for details, see #238.
2. Whereas before, to play an Alda score we fed the text to the parser, which transformed it into Clojure code and then evaluated it (leading to the "Method code too large!" error if too much code was generated), this PR makes it so that the parser transforms the Alda score text directly into the score map that would result from evaluating the Clojure code. So essentially we're skipping a step. Which is weird, because it takes slightly longer when we skip the `eval` step. I guess `eval` must be optimized in some way. Hmm. Anyway, the difference is negligible -- I'll post some benchmarks below -- and it's still significantly faster than the current release, thanks to the parser optimizations.