@daveyarwood 2016-06-20
# Benchmarks

## Before (current release)

Each time a score is played, it is parsed (resulting in Clojure code) and then the resulting Clojure code is evaluated, so the total time for each example here is the time it takes to parse + the time it takes to evaluate.

```
Testing alda.examples-test
Parsing    across_the_sea.alda...                OK (68 ms)
Evaluating across_the_sea.alda...                OK (21 ms)
Parsing    awobmolg.alda...                      OK (402 ms)
Evaluating awobmolg.alda...                      OK (49 ms)
Parsing    bach_cello_suite_no_1.alda...         OK (778 ms)
Evaluating bach_cello_suite_no_1.alda...         OK (104 ms)
Parsing    debussy_quartet.alda...               OK (125 ms)
Evaluating debussy_quartet.alda...               OK (10 ms)
Parsing    entropy.alda...                       OK (119 ms)
Evaluating entropy.alda...                       OK (27 ms)
Parsing    gau.alda...                           OK (193 ms)
Evaluating gau.alda...                           OK (8 ms)
Parsing    hello_world.alda...                   OK (8 ms)
Evaluating hello_world.alda...                   OK (2 ms)
Parsing    key_signature.alda...                 OK (48 ms)
Evaluating key_signature.alda...                 OK (5 ms)
Parsing    multi-poly.alda...                    OK (23 ms)
Evaluating multi-poly.alda...                    OK (6 ms)
Parsing    nesting.alda...                       OK (60 ms)
Evaluating nesting.alda...                       OK (16 ms)
Parsing    overriding-a-global-attribute.alda... OK (31 ms)
Evaluating overriding-a-global-attribute.alda... OK (5 ms)
Parsing    panning.alda...                       OK (75 ms)
Evaluating panning.alda...                       OK (4 ms)
Parsing    percussion.alda...                    OK (109 ms)
Evaluating percussion.alda...                    OK (5 ms)
Parsing    phase.alda...                         OK (17 ms)
Evaluating phase.alda...                         OK (24 ms)
Parsing    poly.alda...                          OK (450 ms)
Evaluating poly.alda...                          OK (45 ms)
Parsing    printing.alda...                      OK (27 ms)
```

## After (this branch)

I adjusted the output of the benchmarks here; **note that the total time for each example is just the 2nd line plus the 3rd line** -- parsing as score and then "realizing" the parsed score. The first line (parsing as code) is just for comparison.

* "Parsing as code" means parsing Alda code and emitting Clojure code (without evaluating it).
* "Parsing as score" means parsing Alda code and transforming it into a sequence of events.
* "Realizing the parsed score" means initializing a new score and reducing through the parsed sequence of events, adding them to the score.

```
Testing alda.examples-test

 across_the_sea.alda
   Parsing as code...        OK (40 ms)
   Parsing as score...       OK (32 ms)
   Realizing parsed score... OK (19 ms)

 awobmolg.alda
   Parsing as code...        OK (218 ms)
   Parsing as score...       OK (222 ms)
   Realizing parsed score... OK (40 ms)

 bach_cello_suite_no_1.alda
   Parsing as code...        OK (395 ms)
   Parsing as score...       OK (497 ms)
   Realizing parsed score... OK (52 ms)

 debussy_quartet.alda
   Parsing as code...        OK (64 ms)
   Parsing as score...       OK (64 ms)
   Realizing parsed score... OK (6 ms)

 entropy.alda
   Parsing as code...        OK (121 ms)
   Parsing as score...       OK (76 ms)
   Realizing parsed score... OK (22 ms)

 gau.alda
   Parsing as code...        OK (68 ms)
   Parsing as score...       OK (64 ms)
   Realizing parsed score... OK (4 ms)

 hello_world.alda
   Parsing as code...        OK (5 ms)
   Parsing as score...       OK (5 ms)
   Realizing parsed score... OK (1 ms)

 key_signature.alda
   Parsing as code...        OK (23 ms)
   Parsing as score...       OK (26 ms)
   Realizing parsed score... OK (4 ms)

 multi-poly.alda
   Parsing as code...        OK (12 ms)
   Parsing as score...       OK (14 ms)
   Realizing parsed score... OK (4 ms)

 nesting.alda
   Parsing as code...        OK (53 ms)
   Parsing as score...       OK (20 ms)
   Realizing parsed score... OK (7 ms)

 overriding-a-global-attribute.alda
   Parsing as code...        OK (14 ms)
   Parsing as score...       OK (14 ms)
   Realizing parsed score... OK (2 ms)

 panning.alda
   Parsing as code...        OK (32 ms)
   Parsing as score...       OK (36 ms)
   Realizing parsed score... OK (2 ms)

 percussion.alda
   Parsing as code...        OK (37 ms)
   Parsing as score...       OK (40 ms)
   Realizing parsed score... OK (3 ms)

 phase.alda
   Parsing as code...        OK (7 ms)
   Parsing as score...       OK (9 ms)
   Realizing parsed score... OK (24 ms)

 poly.alda
   Parsing as code...        OK (179 ms)
   Parsing as score...       OK (179 ms)
   Realizing parsed score... OK (33 ms)

 printing.alda
   Parsing as code...        OK (14 ms)
   Parsing as score...       OK (17 ms)
   Realizing parsed score... OK (1 ms)
```